split in package
